requirements.txt
asgiref==3.4.1
bleach==4.1.0
Django==4.0.1
django-cors-headers==3.11.0
django-summernote==0.8.20.0
djangorestframework==3.13.1
gunicorn==20.1.0
packaging==21.3
pyparsing==3.0.6
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
webencodings==0.5.1
